+++ core-anomaly/shared/src/main/scala/busymachines/pureharm/anomaly/anomaly.scala
@@ -140,6 +140,10 @@ trait AnomalyBase extends Product with Serializable {
   */
 sealed trait StringOrSeqString extends Product with Serializable
 
-final case class StringWrapper private[anomaly] (s: String) extends StringOrSeqString
+final case class StringWrapper private[anomaly] (s: String) extends StringOrSeqString {
+  override val toString: String = s
+}
 
-final case class SeqStringWrapper private[anomaly] (ses: Seq[String]) extends StringOrSeqString
+final case class SeqStringWrapper private[anomaly] (ses: Seq[String]) extends StringOrSeqString {
+  override def toString: String = ses.mkString("[", ",", "]")
+}