+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # 0.2.0
 - add alias for `sprout.Burry` in `PureharmSproutAliases`
-- [pureharm-core-anomaly] now depends on cats. Move `Show[Throwable]` instance to anomaly package. The change is source compatible if you mixed in AnomalyImplicits.
 - remove `Show[New]` instances for any `OldType[Old, New]`, it seriously tripped up type inference
+- [pureharm-core-anomaly] now depends on cats. Move `Show[Throwable]` instance to anomaly package. The change is source compatible if you mixed in AnomalyImplicits.
+- add pretty printed .toString implementation for all Anomalies.
 
 New Scala versions:
 - 3.0.0-RC2